move APP component to correct folder, add book as a prop for Book component and erase debugging statements
